+++ src/main.cpp
@@ -5,6 +5,7 @@
 #include <platform/assert.h>
 #include <platform/commands.h>
 #include <platform/file.h>
+#include <platform/font.h>
 #include <platform/image.h>
 #include <platform/input/input.h>
 #include <platform/input/timing.h>
@@ -39,32 +40,11 @@ using LoadLibraryError = platform::LoadLibraryError;
 using Renderer = platform::Renderer;
 using ShaderProgram = platform::ShaderProgram;
 using ShaderProgramError = platform::ShaderProgramError;
-using Vertex = platform::Vertex;
-using VertexSection = platform::VertexSection;
 
 struct FullscreenState {
 	bool is_fullscreen = false;
 	int last_windowed_x = 0;
 	int last_windowed_y = 0;
-};
-
-struct Glyph {
-	glm::ivec2 atlas_pos;
-	glm::ivec2 size;
-	glm::ivec2 bearing;
-	int advance;
-};
-
-struct Font {
-	static constexpr size_t NUM_GLYPHS = 127;
-	Glyph glyphs[NUM_GLYPHS]; // indexed with printable ascii values
-	platform::Texture atlas;
-};
-
-struct RGB {
-	uint8_t r;
-	uint8_t g;
-	uint8_t b;
 };
 
 const char* LIBRARY_NAME = "GameEngine2024";
@@ -169,78 +149,6 @@ void start_imgui_frame() {
 	ImGui::NewFrame();
 }
 
-// TODO make FT_library a global var to make API cleaner
-// no sensible way to pass FT_library around with abstraction leakage towards engine
-Font add_font(FT_Library ft, const char* font_path, uint8_t font_size) {
-	Font font;
-
-	/* Load font */
-	FT_Face face;
-	if (FT_Error error = FT_New_Face(ft, font_path, 0, &face); error != FT_Err_Ok) {
-		// FIXME return an error here instead of aborting
-		ABORT("FT_New_Face(\"%s\") failed: %s", font_path, FT_Error_String(error));
-	}
-
-	/* Set font size */
-	int font_upscale = 64;
-	FT_Set_Char_Size(face, 0, font_size * font_upscale, 96, 96);
-
-	/* Calculate atlas dimensions to be a square */
-	uint32_t glyph_height = (1 + (face->size->metrics.height / font_upscale));
-	uint32_t glyph_width = glyph_height / 2; // assume 2:1 ratio
-	uint32_t columns = (uint32_t)roundf(sqrtf((float)Font::NUM_GLYPHS * (float)glyph_height / (float)glyph_width));
-	uint32_t rows = (uint32_t)roundf((float)Font::NUM_GLYPHS / (float)columns);
-	uint32_t texture_width = columns * glyph_width;
-	uint32_t texture_height = rows * glyph_height;
-
-	/* Compute glyphs */
-	std::vector<uint8_t> glyph_pixels = std::vector<uint8_t>(texture_width * texture_height);
-	glm::ivec2 pen = { 0, 1 };
-	for (int i = '!'; i < Font::NUM_GLYPHS; i++) {
-		// load character
-		FT_Load_Char(face, i, FT_LOAD_RENDER | FT_LOAD_FORCE_AUTOHINT | FT_LOAD_TARGET_LIGHT);
-		FT_Bitmap* bmp = &face->glyph->bitmap;
-
-		// render current glyph
-		for (uint32_t row = 0; row < bmp->rows; row++) {
-			for (uint32_t col = 0; col < bmp->width; col++) {
-				uint32_t x = pen.x + col;
-				uint32_t y = pen.y + row;
-				glyph_pixels[y * texture_width + x] = bmp->buffer[row * bmp->pitch + col];
-			}
-		}
-
-		// save glyph info
-		font.glyphs[i].atlas_pos = pen;
-		font.glyphs[i].size = { bmp->width, bmp->rows };
-		font.glyphs[i].bearing = { face->glyph->bitmap_left, face->glyph->bitmap_top };
-		font.glyphs[i].advance = face->glyph->advance.x / font_upscale;
-
-		// move pen
-		pen.x += bmp->width + 1;
-		if (pen.x + bmp->width >= texture_width) {
-			pen.x = 0;
-			pen.y += face->size->metrics.height / font_upscale + 2;
-		}
-	}
-	FT_Done_Face(face);
-
-	/* Generate glyph texture */
-	std::vector<RGB> glyph_rgb = std::vector<RGB>(texture_width * texture_height);
-	for (uint32_t y = 0; y < texture_height; y++) {
-		uint32_t inv_y = (texture_height - 1) - y;
-		for (uint32_t x = 0; x < texture_width; x++) {
-			glyph_rgb[inv_y * texture_width + x].r |= glyph_pixels[y * texture_width + x];
-			glyph_rgb[inv_y * texture_width + x].g |= glyph_pixels[y * texture_width + x];
-			glyph_rgb[inv_y * texture_width + x].b |= glyph_pixels[y * texture_width + x];
-		}
-	}
-
-	font.atlas = platform::add_texture((uint8_t*)glyph_rgb.data(), texture_width, texture_height);
-
-	return font;
-}
-
 int main(int /* argc */, char** /* args */) {
 	glm::ivec2 resolution = { 800, 600 };
 
@@ -306,7 +214,7 @@ int main(int /* argc */, char** /* args */) {
 	FullscreenState fullscreen_state;
 
 	// add test font
-	Font arial_font = add_font(ft, "C:/windows/Fonts/Arial.ttf", 16);
+	platform::Font arial_font = platform::add_font(ft, "C:/windows/Fonts/Arial.ttf", 16);
 
 	engine.initialize(&state);
 	while (!quit) {
@@ -347,7 +255,7 @@ int main(int /* argc */, char** /* args */) {
 				renderer.draw_texture({ 0.0f, 0.0f }, { arial_font.atlas.width, arial_font.atlas.height }, arial_font.atlas);
 				// render glyph box
 				auto draw_glyph_box = [&](uint8_t ch) {
-					const Glyph& glyph = arial_font.glyphs[ch];
+					const platform::Glyph& glyph = arial_font.glyphs[ch];
 					glm::vec2 top_left = glyph.atlas_pos;
 					glm::vec2 bottom_right = glyph.atlas_pos + glyph.size;
 					renderer.draw_rect(top_left, bottom_right, glm::vec4 { 0.0f, 1.0f, 0.0f, 1.0f });
